remove lightgreen from the visual feedback
mistake numbers are important to distinguish when they are low,
the most common number is zero. there is basically no difference
between 5 and 6 mistakes in a single phrase.

all that means that we should very clearly distinguish between zero
mistakes and one mistake.

this commit removes the lightgreen (color of one mistake), which is very
close to the green of zero mistakes. thus one mistakes becomes limegreen
(formerly the color of two mistakes).

and the hint now shows after only five mistakes not six.