improve selectors
aayaat fields now are selectors like the suar; that should generally
be easier (especially on phones and tablets) and less error-prone.

these fields support searching by keyboard (eg, typing 9 to select
the 9th aayah, or typing 1 2 3 in quick succesion to select the
123rd aayah.

that works normally if your keyboard layout produces Eastern Arabic
numerals (٠١٢٣٤٥٦٧٨٩), but most keyboards don't, even the Arabic ones.

I managed to get it to work with ASCII numerals (0123456789), but it
works only when the field is focused but closed. if it's open it
doesn't work; use up & down arrows, or use your mouse.

this commit is largely equivalent to recite's `afc7f65` commit
(<https://github.com/noureddin/recite/>).